+++ nanoFramework.Tools.Hex2Dfu.sln
@@ -3,24 +3,34 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.30907.101
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "hex2dfu", "hex2dfu.csproj", "{85CA39C2-5404-4E03-8CB5-2E098ABBC02C}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "hex2dfu", "source\hex2dfu.csproj", "{85CA39C2-5404-4E03-8CB5-2E098ABBC02C}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Hex2BinTest", ".\Hex2BinTest\Hex2BinTest.csproj", "{0590D59E-F2F4-4075-BB72-A69DD5AF1BFA}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Hex2BinTest", "source\Hex2BinTest\Hex2BinTest.csproj", "{0590D59E-F2F4-4075-BB72-A69DD5AF1BFA}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
 		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{85CA39C2-5404-4E03-8CB5-2E098ABBC02C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{85CA39C2-5404-4E03-8CB5-2E098ABBC02C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{85CA39C2-5404-4E03-8CB5-2E098ABBC02C}.Debug|x64.ActiveCfg = Debug|x64
+		{85CA39C2-5404-4E03-8CB5-2E098ABBC02C}.Debug|x64.Build.0 = Debug|x64
 		{85CA39C2-5404-4E03-8CB5-2E098ABBC02C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{85CA39C2-5404-4E03-8CB5-2E098ABBC02C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{85CA39C2-5404-4E03-8CB5-2E098ABBC02C}.Release|x64.ActiveCfg = Release|x64
+		{85CA39C2-5404-4E03-8CB5-2E098ABBC02C}.Release|x64.Build.0 = Release|x64
 		{0590D59E-F2F4-4075-BB72-A69DD5AF1BFA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{0590D59E-F2F4-4075-BB72-A69DD5AF1BFA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0590D59E-F2F4-4075-BB72-A69DD5AF1BFA}.Debug|x64.ActiveCfg = Debug|x64
+		{0590D59E-F2F4-4075-BB72-A69DD5AF1BFA}.Debug|x64.Build.0 = Debug|x64
 		{0590D59E-F2F4-4075-BB72-A69DD5AF1BFA}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{0590D59E-F2F4-4075-BB72-A69DD5AF1BFA}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0590D59E-F2F4-4075-BB72-A69DD5AF1BFA}.Release|x64.ActiveCfg = Release|x64
+		{0590D59E-F2F4-4075-BB72-A69DD5AF1BFA}.Release|x64.Build.0 = Release|x64
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
