+++ source/BinaryFileInfo.cs
