+++ source/Hex2Dfu.cs
@@ -1,244 +1,158 @@
 ﻿//
-// Copyright (c) 2017 The nanoFramework project contributors
-// Portions Copyright (c) COPYRIGHT 2015 STMicroelectronics
+// Copyright (c) 2021 The nanoFramework project contributors
 // See LICENSE file in the project root for full license information.
 //
 
 using System;
+using System.Buffers.Binary;
 using System.Collections.Generic;
 using System.IO;
-using System.Linq;
-using System.Runtime.InteropServices;
 
 namespace nanoFramework.Tools
 {
+    /// <summary>
+    /// HEX to DFU and BIN to DFU converter class
+    /// </summary>
     public class Hex2Dfu
     {
-
-        #region constants from STDFUFiles
+        private const ushort DefaultSTMVid = 0x0483;
+        private const ushort DefafultSTMPid = 0xDF11;
+        private const ushort DefaultFwVersion = 0x2200;
 
         /// <summary>
-        /// No error.
+        /// Concert a HEX file into a DFU file, save the intermediate BIN file
         /// </summary>
-        const uint STDFUFILES_NOERROR = 0x12340000;
-
-        #endregion
-
-
-        //a block of data to be written
-        [StructLayout(LayoutKind.Sequential, Pack = 1)]
-        unsafe public struct STDFUFILES_DfuImageElement
+        /// <param name="hexFile">The original HEX file name</param>
+        /// <param name="dfuName">The target DFU file name</param>
+        /// <param name="vid">The Vendor ID</param>
+        /// <param name="pid">The Product ID</param>
+        /// <param name="fwVersion">The Firmware Version</param>
+        /// <returns>True if success</returns>
+        public static bool CreateDfuFile(string hexFile, string dfuName, ushort vid = DefaultSTMVid, ushort pid = DefafultSTMPid, ushort fwVersion = DefaultFwVersion)
         {
-            public UInt32 dwAddress;
-            public UInt32 dwDataLength;
-            public IntPtr Data;
-        }
-
-
-        #region imports from STDFUFiles.dll
-
-        // from dumpbin
-        // 1    0 00004480 STDFUFILES_AppendImageToDFUFile
-        // 2    1 000044E0 STDFUFILES_CloseDFUFile
-        // 3    2 00004550 STDFUFILES_CreateImage
-        // 4    3 000045B0 STDFUFILES_CreateImageFromMapping
-        // 5    4 00004610 STDFUFILES_CreateNewDFUFile
-        // 6    5 00004680 STDFUFILES_DestroyImage
-        // 7    6 00004700 STDFUFILES_DestroyImageElement
-        // 8    7 00004770 STDFUFILES_DuplicateImage
-        // 9    8 00004820 STDFUFILES_FilterImageForOperation
-        //10    9 000048A0 STDFUFILES_GetImageAlternate
-        //11    A 00004900 STDFUFILES_GetImageElement
-        //12    B 00004970 STDFUFILES_GetImageName
-        //13    C 000049E0 STDFUFILES_GetImageNbElement
-        //14    D 00004A40 STDFUFILES_GetImageSize
-        //15    E 00004A90 STDFUFILES_ImageFromFile
-        //16    F 00004B10 STDFUFILES_ImageToFile
-        //17   10 00004B80 STDFUFILES_OpenExistingDFUFile
-        //18   11 00004C20 STDFUFILES_ReadImageFromDFUFile
-        //19   12 00004CA0 STDFUFILES_SetImageElement
-        //20   13 00004D20 STDFUFILES_SetImageName
-
-        [DllImport("STDFUFiles.DLL", EntryPoint = "STDFUFILES_AppendImageToDFUFile", CharSet = CharSet.Auto)]
-        public static extern UInt32 STDFUFILES_AppendImageToDFUFile(IntPtr handle, IntPtr image);
-
-        [DllImport("STDFUFiles.DLL", EntryPoint = "STDFUFILES_CloseDFUFile", CharSet = CharSet.Auto)]
-        public static extern UInt32 STDFUFILES_CloseDFUFile(IntPtr handle);
-
-        [DllImport("STDFUFiles.DLL", EntryPoint = "STDFUFILES_CreateNewDFUFile", CharSet = CharSet.Auto)]
-        public static extern UInt32 STDFUFILES_CreateNewDFUFile([MarshalAs(UnmanagedType.LPStr)]String szDevicePath, ref IntPtr handle, UInt16 Vid, UInt16 Pid, UInt16 Bcd);
-
-        [DllImport("STDFUFiles.DLL", EntryPoint = "STDFUFILES_CreateImage", CharSet = CharSet.Auto)]
-        public static extern UInt32 STDFUFILES_CreateImage(ref IntPtr image, byte nAlternate);
-
-        [DllImport("STDFUFiles.DLL", EntryPoint = "STDFUFILES_DestroyImage", CharSet = CharSet.Auto)]
-        public static extern UInt32 STDFUFILES_DestroyImage(ref IntPtr handle);
-
-        [DllImport("STDFUFiles.DLL", EntryPoint = "STDFUFILES_ImageFromFile", CharSet = CharSet.Auto)]
-        public static extern UInt32 STDFUFILES_ImageFromFile([MarshalAs(UnmanagedType.LPStr)]String szDevicePath, ref IntPtr image, byte nAlternate);
-
-        [DllImport("STDFUFiles.DLL", EntryPoint = "STDFUFILES_SetImageElement", CharSet = CharSet.Auto)]
-        public static extern UInt32 STDFUFILES_SetImageElement(IntPtr handle, UInt32 dwRank, bool bInsert, [MarshalAs(UnmanagedType.Struct)]STDFUFILES_DfuImageElement Element);
-
-        [DllImport("STDFUFiles.DLL", EntryPoint = "STDFUFILES_SetImageName", CharSet = CharSet.Auto)]
-        public static extern UInt32 STDFUFILES_SetImageName(IntPtr image, [MarshalAs(UnmanagedType.LPStr)]String pPathFile);
-
-        #endregion
-
-        const UInt16 defaultSTMVid = 0x0483;
-        const UInt16 defafultSTMPid = 0xDF11;
-        const UInt16 defaultFwVersion = 0x2200;
-
-        public static bool CreateDfuFile(string hexFile, string dfuName, UInt16 vid = defaultSTMVid, UInt16 pid = defafultSTMPid, UInt16 fwVersion = defaultFwVersion)
-        {
-            IntPtr dfuFileHandle = (IntPtr)0;
-            IntPtr imageFileHandle = (IntPtr)0;
-
-            // start creating a new DFU file for output
-            var retCode = STDFUFILES_CreateNewDFUFile(dfuName, ref dfuFileHandle, vid, pid, fwVersion);
-
-            if (retCode == STDFUFILES_NOERROR)
+            using StreamReader reader = new StreamReader(hexFile);
+            using MemoryStream binDestination = new MemoryStream();
+            string hexLine = string.Empty;
+            long offset = 0;
+            long maxSize = 0;
+            Console.WriteLine($"Converting HEX file {hexFile} to BIN format.");
+            while (reader.Peek() >= 0)
             {
-
-                // get image from HEX file
-                retCode = STDFUFILES_ImageFromFile(hexFile, ref imageFileHandle, 0);
-
-                if (retCode == STDFUFILES_NOERROR)
+                hexLine = reader.ReadLine();
+                HexFormat hex = new HexFormat(hexLine);
+                if (!hex.IsValidRecord)
                 {
-                    // add image of HEX file
-                    retCode = STDFUFILES_AppendImageToDFUFile(dfuFileHandle, imageFileHandle);
+                    Console.WriteLine();
+                    Console.WriteLine($"ERROR: Invalid HEX record");
+                    Console.WriteLine();
+                    throw new Exception($"Error reading hex file, invalid file or CRC error");
+                }
 
-                    if (retCode != STDFUFILES_NOERROR)
-                    {
-                        // error adding this file
-                        Console.WriteLine();
-                        Console.WriteLine($"ERROR: adding {hexFile}");
-                        Console.WriteLine();
-
-                        return false;
-                    }
-
-                    Console.WriteLine($"Adding image for {hexFile}");
+                if (hex.HexFieldType == HexFieldType.Data)
+                {
+                    binDestination.Seek(offset + hex.Address, SeekOrigin.Begin);
+                    maxSize = maxSize < offset + hex.Address ? offset + hex.Address : maxSize;
+                    binDestination.Write(hex.Data);
+                }
+                else if (hex.HexFieldType == HexFieldType.ExtendedAddress)
+                {
+                    offset = BinaryPrimitives.ReadUInt16BigEndian(hex.Data) << 4;
+                }
+                else if (hex.HexFieldType == HexFieldType.EndOfFile)
+                {
+                    break;
                 }
             }
 
-            // image file added, close DFU file
-            STDFUFILES_CloseDFUFile(dfuFileHandle);
+            // Close the reader
+            reader.Close();
+            reader.Dispose();
+            // Save the file result and close the memory stream and the file
+            using FileStream fs = new FileStream(hexFile + ".bin", FileMode.Create);
+            Span<byte> binFile = binDestination.GetBuffer().AsSpan(0, (int)maxSize);
+            binDestination.Close();
+            binDestination.Dispose();
+            fs.Write(binFile);
+            fs.Close();
+            fs.Dispose();
 
+            DfuFile dfuFile = new();
+            DfuImage dfuImage = new();
+            dfuImage.TargetPrefix.TargetName = GetCleanName(dfuName);
+            ImageElement imageElement = new();
+            imageElement.Data = binFile.ToArray();
+            dfuImage.ImageElements.Add(imageElement);
+            dfuFile.DfuImages.Add(dfuImage);
+            dfuFile.DfuSuffix.FirmwareVersion = fwVersion;
+            dfuFile.DfuSuffix.ProductId = pid;
+            dfuFile.DfuSuffix.VersionId = vid;
+            var ser = dfuFile.Serialize();
+            using FileStream fdest = new FileStream(dfuName, FileMode.Create);
+            fdest.Write(ser);
+            fdest.Close();
+            fdest.Dispose();
             Console.WriteLine();
             Console.WriteLine($"DFU generated: {dfuName}");
             Console.WriteLine($"Vendor ID: {vid.ToString("X4")}");
             Console.WriteLine($"Product ID: {pid.ToString("X4")}");
             Console.WriteLine($"Version: {fwVersion.ToString("X4")}");
             Console.WriteLine();
-
-            // clean-up
-            if (retCode == STDFUFILES_NOERROR)
-            {
-                // destroy image
-                STDFUFILES_DestroyImage(ref imageFileHandle);
-            }
-
             return true;
         }
 
-        public static bool CreateDfuFile(List<BinaryFileInfo> binFiles, string dfuName, UInt16 vid = defaultSTMVid, UInt16 pid = defafultSTMPid, UInt16 fwVersion = defaultFwVersion)
+        /// <summary>
+        /// Convert multiple BIN files with specific target address to a DFU file
+        /// </summary>
+        /// <param name="binFiles">A list of Binary File Information</param>
+        /// <param name="dfuName">The target DFU file name</param>
+        /// <param name="vid">The Vendor ID</param>
+        /// <param name="pid">The Product ID</param>
+        /// <param name="fwVersion">The Firmware Version</param>
+        /// <returns></returns>
+        public static bool CreateDfuFile(List<BinaryFileInfo> binFiles, string dfuName, ushort vid = DefaultSTMVid, ushort pid = DefafultSTMPid, ushort fwVersion = DefaultFwVersion)
         {
-            IntPtr dfuFileHandle = (IntPtr)0;
-            IntPtr imageFileHandle = (IntPtr)0;
-
-            // start creating a new DFU file for output
-            var retCode = STDFUFILES_CreateNewDFUFile(dfuName, ref dfuFileHandle, vid, pid, fwVersion);
-
-            if (retCode == STDFUFILES_NOERROR)
+            DfuFile dfuFile = new();
+            DfuImage dfuImage = new();
+            dfuImage.TargetPrefix.TargetName = "nanoFramework";
+            foreach (var binFile in binFiles)
             {
-                retCode = STDFUFILES_CreateImage(ref imageFileHandle, 0);
-
-                retCode = STDFUFILES_SetImageName(imageFileHandle, "nanoFramework");
-                uint fileCounter = 0;
-
-                // loop through collection of bin files and add them
-                foreach (BinaryFileInfo file in binFiles)
+                using FileStream fs = new(binFile.FileName, FileMode.Open);
+                ImageElement imageElement = new();
+                imageElement.Data = new byte[fs.Length];
+                imageElement.ElementAddress = binFile.Address;
+                if (fs.Read(imageElement.Data) <= 0)
                 {
-                    byte[] fileData = File.ReadAllBytes(file.FileName);
-
-                    // get required memory size for byte array
-                    int size = Marshal.SizeOf(fileData[0]) * fileData.Length;
-
-                    STDFUFILES_DfuImageElement element = new STDFUFILES_DfuImageElement();
-                    element.dwAddress = file.Address;
-                    element.dwDataLength = (uint)fileData.Length;
-                    
-                    // allocate memory from the unmanaged memory
-                    element.Data = Marshal.AllocHGlobal(size);
-
-                    // copy the byte array to the struct
-                    Marshal.Copy(fileData, 0, element.Data, fileData.Length);
-
-                    // get image from HEX file
-                    retCode = STDFUFILES_SetImageElement(imageFileHandle, fileCounter++, true, element);
-
-                    // free unmanaged memory
-                    Marshal.FreeHGlobal(element.Data);
-
-                    if (retCode != STDFUFILES_NOERROR)
-                    {
-                        // error adding this file
-                        Console.WriteLine();
-                        Console.WriteLine($"ERROR: adding {file.FileName}");
-                        Console.WriteLine();
-
-                        return false;
-                    }
-
-                    Console.WriteLine($"Adding file to image: {file.FileName}");
-                }
-
-                // add image to DFU file
-                retCode = STDFUFILES_AppendImageToDFUFile(dfuFileHandle, imageFileHandle);
-
-                if (retCode != STDFUFILES_NOERROR)
-                {
-                    // error adding this file
                     Console.WriteLine();
-                    Console.WriteLine($"ERROR: adding image to DFU file");
+                    Console.WriteLine($"ERROR: adding {binFile.FileName}");
                     Console.WriteLine();
-
                     return false;
                 }
 
-                // image file added, close DFU file
-                STDFUFILES_CloseDFUFile(dfuFileHandle);
-
-                Console.WriteLine();
-                Console.WriteLine($"DFU generated: {dfuName}");
-                Console.WriteLine($"Vendor ID: {vid.ToString("X4")}");
-                Console.WriteLine($"Product ID: {pid.ToString("X4")}");
-                Console.WriteLine($"Version: {fwVersion.ToString("X4")}");
-                Console.WriteLine();
-
-                // clean-up
-                if (retCode == STDFUFILES_NOERROR)
-                {
-                    // destroy image
-                    STDFUFILES_DestroyImage(ref imageFileHandle);
-                }
+                Console.WriteLine($"Adding file to image: {binFile.FileName}");
+                dfuImage.ImageElements.Add(imageElement);
             }
 
+            dfuFile.DfuImages.Add(dfuImage);
+            dfuFile.DfuSuffix.FirmwareVersion = fwVersion;
+            dfuFile.DfuSuffix.ProductId = pid;
+            dfuFile.DfuSuffix.VersionId = vid;
+            var ser = dfuFile.Serialize();
+            using FileStream fdest = new FileStream(dfuName, FileMode.Create);
+            fdest.Write(ser);
+            fdest.Close();
+            fdest.Dispose();
+            Console.WriteLine();
+            Console.WriteLine($"DFU generated: {dfuName}");
+            Console.WriteLine($"Vendor ID: {vid.ToString("X4")}");
+            Console.WriteLine($"Product ID: {pid.ToString("X4")}");
+            Console.WriteLine($"Version: {fwVersion.ToString("X4")}");
+            Console.WriteLine();
             return true;
         }
 
-    }
-
-    public class BinaryFileInfo
-    {
-        public string FileName { get; private set; }
-        public uint Address { get; private set; }
-
-        public BinaryFileInfo(string fileName, uint address)
+        private static string GetCleanName(string fileName)
         {
-            FileName = fileName;
-            Address = address;
+            return Path.GetFileNameWithoutExtension(fileName);
         }
     }
 }
