--- source/Properties/AssemblyInfo.cs
@@ -1,36 +1,0 @@
-﻿using System.Reflection;
-using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
-
-// General Information about an assembly is controlled through the following
-// set of attributes. Change these attribute values to modify the information
-// associated with an assembly.
-[assembly: AssemblyTitle("nanoFramework Tools Hex2Dfu")]
-[assembly: AssemblyDescription("Utility to convert HEX files into DFU target image")]
-[assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("nanoFramework project contributors")]
-[assembly: AssemblyProduct("nanoFramework Tools Hex2Dfu")]
-[assembly: AssemblyCopyright("Copyright ©  2017 nanoFramework project contributors")]
-[assembly: AssemblyTrademark("")]
-[assembly: AssemblyCulture("")]
-
-// Setting ComVisible to false makes the types in this assembly not visible
-// to COM components.  If you need to access a type in this assembly from
-// COM, set the ComVisible attribute to true on that type.
-[assembly: ComVisible(false)]
-
-// The following GUID is for the ID of the typelib if this project is exposed to COM
-[assembly: Guid("378070ef-3572-47f4-a854-3d6c5da3eefa")]
-
-// Version information for an assembly consists of the following four values:
-//
-//      Major Version
-//      Minor Version
-//      Build Number
-//      Revision
-//
-// You can specify all the values or you can default the Build and Revision Numbers
-// by using the '*' as shown below:
-// [assembly: AssemblyVersion("1.0.*")]
-
-
