--- source/nanoFramework.Tools.Hex2Dfu.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.26730.15
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "nanoFramework.Tools.Hex2Dfu", "nanoFramework.Tools.Hex2Dfu.csproj", "{378070EF-3572-47F4-A854-3D6C5DA3EEFA}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{378070EF-3572-47F4-A854-3D6C5DA3EEFA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{378070EF-3572-47F4-A854-3D6C5DA3EEFA}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{378070EF-3572-47F4-A854-3D6C5DA3EEFA}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{378070EF-3572-47F4-A854-3D6C5DA3EEFA}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {0CFE97F1-0501-4C1F-8CC5-11043784ECC5}
-	EndGlobalSection
-EndGlobal
