--- source/packages.config
@@ -1,6 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Costura.Fody" version="3.3.3" targetFramework="net46" developmentDependency="true" />
-  <package id="Fody" version="4.2.1" targetFramework="net46" developmentDependency="true" />
-  <package id="Nerdbank.GitVersioning" version="3.0.19-beta" targetFramework="net46" developmentDependency="true" />
-</packages>
