--- source/version.json
@@ -1,16 +1,0 @@
-{
-  "$schema": "https://raw.githubusercontent.com/AArnott/Nerdbank.GitVersioning/master/src/NerdBank.GitVersioning/version.schema.json",
-  "version": "1.3",
-  "assemblyVersion": {
-    "precision": "build"
-  },
-  "semVer1NumericIdentifierPadding": 3,
-  "publicReleaseRefSpec": [
-    "^refs/heads/master$",
-    "^refs/heads/v\\d+(?:\\.\\d+)?$"
-  ],
-  "cloudBuild": {
-    "setVersionVariables": true,
-    "setAllVariables": true
-  }
-}
